backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
langdetect==1.0.9
python-dotenv==1.0.0
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6